+++ core/handle/textHandle.py
@@ -2,7 +2,7 @@ from config.logger import setup_logging
 import json
 from core.handle.abortHandle import handleAbortMessage
 from core.handle.helloHandle import handleHelloMessage
-from core.handle.audioHandle import startToChat
+from core.handle.receiveAudioHandle import startToChat
 from core.handle.iotHandle import handleIotDescriptors
 
 TAG = __name__
@@ -39,7 +39,6 @@ async def handleTextMessage(conn, message):
                     await startToChat(conn, msg_json["text"])
         elif msg_json["type"] == "iot":
             if "descriptors" in msg_json:
-                pass
-                # await handleIotDescriptors(conn, msg_json["descriptors"])
+                await handleIotDescriptors(conn, msg_json["descriptors"])
     except json.JSONDecodeError:
         await conn.websocket.send(message)
